# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pyKorf",
    "version": "0.18.0"
  },
  "paths": {
...
    "/api/settings/apply": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Apply Settings",
        "description": "Apply selected global settings to the model.",
        "operationId": "applySettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyGlobalSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsApplyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/center-layout": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Center Layout",
        "description": "Center all elements on the page.",
        "operationId": "centerLayout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CenterLayoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/settings/snap-orthogonal": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Snap Orthogonal",
        "description": "Snap near-orthogonal connections and align to grid.",
        "operationId": "snapOrthogonal",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapOrthogonalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CenterLayoutResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/report/generate": {
      "post": {
        "tags": [
          "report"
        ],
        "summary": "Generate Report",
        "description": "Generate a single-model Excel report.",
        "operationId": "generateReport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/report/export": {
      "post": {
        "tags": [
          "report"
        ],
        "summary": "Export Excel",
        "description": "Export model to Excel file.",
        "operationId": "exportReport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/report/import": {
      "post": {
        "tags": [
          "report"
        ],
        "summary": "Import Excel",
        "description": "Import model parameters from Excel file.",
        "operationId": "importReport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/report/batch": {
      "post": {
        "tags": [
          "report"
        ],
        "summary": "Batch Report",
        "description": "Generate batch report across multiple KDF files in a folder.",
        "operationId": "batchReport",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchReportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplyGlobalSettingsRequest": {
        "properties": {
          "setting_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Setting Ids"
          },
          "dp_margin": {
            "type": "number",
            "title": "Dp Margin",
            "default": 1.25
          },
          "shutoff_margin": {
            "type": "number",
            "title": "Shutoff Margin",
            "default": 1.2
          },
          "min_pump_elevation": {
            "type": "number",
            "title": "Min Pump Elevation",
            "default": 0.5
          }
        },
        "type": "object",
        "required": [
          "setting_ids"
        ],
        "title": "ApplyGlobalSettingsRequest"
      },
...
      "BatchReportRequest": {
        "properties": {
          "batch_folder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Folder"
          },
          "single_report": {
            "type": "boolean",
            "title": "Single Report",
            "default": false
          }
        },
        "type": "object",
        "title": "BatchReportRequest"
      },
...
      "CenterLayoutResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "CenterLayoutResponse"
      },
...
      "EmptyRequest": {
        "properties": {},
        "type": "object",
        "title": "EmptyRequest",
        "description": "Explicit empty request payload for body-less POST endpoints."
      },
      "ExportRequest": {
        "properties": {
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          }
        },
        "type": "object",
        "title": "ExportRequest"
      },
      "GenerateReportRequest": {
        "properties": {
          "report_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Path"
          }
        },
        "type": "object",
        "title": "GenerateReportRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportRequest": {
        "properties": {
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          }
        },
        "type": "object",
        "title": "ImportRequest"
      },
...
      "ReportResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/StatusMessage"
            },
            "type": "array",
            "title": "Messages",
            "default": []
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ReportResponse"
      },
...
      "SettingsApplyResponse": {
        "properties": {
          "results": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Results",
            "default": {}
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "title": "SettingsApplyResponse"
      },
...
      "SnapOrthogonalRequest": {
        "properties": {
          "threshold_deg": {
            "type": "number",
            "title": "Threshold Deg",
            "default": 10.0
          },
          "grid_size": {
            "type": "number",
            "title": "Grid Size",
            "default": 500.0
          }
        },
        "type": "object",
        "title": "SnapOrthogonalRequest"
      },
      "StatusMessage": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "type",
          "message"
        ],
        "title": "StatusMessage"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}